Yopa exercise
A notes application that allows users to create, read and delete notes.

Code used:

HTML, CSS, Sass, Bootstrap, Jquery, and Javascript.

The site is fully responsive.
Create notes by typing into the text field and pressing add note.
Click the plus symbol on the new note to add to important notes.
Click the cross to remove notes.
Press add note without user input and error message is displayed.
Login modal created on click of login button.

The login section displays the fields that would be used to log in, and
would be linked up to a database to allow users to log in and access
their notes.